+++ apworld/brotato/options.py
@@ -112,6 +112,31 @@ class WavesPerCheck(Range):
     display_name = "Waves Per Check"
 
 
+class NumWaveCaps(Choice):
+    """The number of caps to put on the accessible waves.
+
+    You will not be able to play beyond the maximum wave cap until you find additional
+    Progressive Wave Cap Increase items.
+
+    With one cap, all waves are accessible from the start.
+
+    With two caps, waves 1-10 are accessible from the start, and waves 11-20 require one
+    Progressive Wave Cap Increase.
+
+    With four caps: waves 1-5 are accessible from the start, waves 6-10 require one
+    Progressive Wave Cap Increase, waves 11-15 require two, waves 16-20 require 3.
+    """
+
+    option_one = 0
+    option_two = 2
+    option_four = 4
+    option_ten = 10
+    option_twenty = 20
+
+    default = 2
+    display_name = "Number of Progressive Wave Caps"
+
+
 class GoldRewardMode(Choice):
     """Chooses how gold rewards are given.
 
@@ -485,6 +510,7 @@ class BrotatoOptions(PerGameCommonOptions, DeathLinkMixin):
     starting_characters: StartingCharacters
     include_base_game_characters: IncludeBaseGameCharacters
     waves_per_drop: WavesPerCheck
+    num_wave_caps: NumWaveCaps
     gold_reward_mode: GoldRewardMode
     xp_reward_mode: XpRewardMode
     enable_enemy_xp: EnableEnemyXp