+++ apworld/brotato/regions.py
@@ -15,10 +15,7 @@ from .constants import (
 )
 from .locations import BrotatoCommonCrateLocation, BrotatoLegendaryCrateLocation, BrotatoLocation, location_table
 from .loot_crates import BrotatoLootCrateGroup
-from .rules import (
-    create_has_character_rule,
-    create_has_run_wins_rule,
-)
+from .rules import create_can_reach_wave_rule, create_has_character_rule, create_has_run_wins_rule
 
 RegionFactory = Callable[[str], Region]
 
@@ -27,6 +24,7 @@ def create_regions(
     region_factory: RegionFactory,
     characters: list[str],
     waves_with_checks: list[int],
+    wave_access: dict[int, int],
     common_loot_crate_groups: list[BrotatoLootCrateGroup],
     legendary_loot_crate_groups: list[BrotatoLootCrateGroup],
 ) -> list[Region]:
@@ -50,7 +48,7 @@ def create_regions(
             regions.append(loot_crate_group_region)
 
     for char in characters:
-        character_region = create_character_region(region_factory, char, waves_with_checks)
+        character_region = create_character_region(region_factory, char, waves_with_checks, wave_access)
         has_character_rule = create_has_character_rule(character_region.player, char)
         menu_region.connect(character_region, f"Start Game ({char})", rule=has_character_rule)
         regions.append(character_region)
@@ -58,20 +56,44 @@ def create_regions(
     return regions
 
 
-def create_character_region(create_region: RegionFactory, character: str, waves_with_checks: list[int]) -> Region:
+def create_character_region(
+    create_region: RegionFactory,
+    character: str,
+    waves_with_checks: list[int],
+    wave_access: dict[int, int],
+) -> Region:
     character_region: Region = create_region(CHARACTER_REGION_TEMPLATE.format(char=character))
     run_complete_location_name = RUN_COMPLETE_LOCATION_TEMPLATE.format(char=character)
-    region_locations: dict[str, int | None] = {
-        run_complete_location_name: location_table[run_complete_location_name].id
-    }
+    run_complete_location_id = location_table[run_complete_location_name].id
+    run_complete_location = BrotatoLocation(
+        character_region.player,
+        name=run_complete_location_name,
+        address=run_complete_location_id,
+        parent=character_region,
+    )
+    # All wave cap increases are needed to reach the final wave.
+    run_complete_location.access_rule = create_can_reach_wave_rule(character_region.player, wave_access[NUM_WAVES])
+    character_region.locations.append(run_complete_location)
 
     for wave in waves_with_checks:
         if wave not in range(1, NUM_WAVES + 1):
             raise ValueError(f"Invalid wave number {wave}.")
         wave_complete_location_name = WAVE_COMPLETE_LOCATION_TEMPLATE.format(wave=wave, char=character)
-        region_locations[wave_complete_location_name] = location_table[wave_complete_location_name].id
+        wave_complete_location_id = location_table[wave_complete_location_name].id
 
-    character_region.add_locations(region_locations, BrotatoLocation)
+        wave_complete_location = BrotatoLocation(
+            character_region.player,
+            name=wave_complete_location_name,
+            address=wave_complete_location_id,
+            parent=character_region,
+        )
+
+        num_wave_cap_items_needed_for_wave = wave_access[wave]
+        wave_complete_location.access_rule = create_can_reach_wave_rule(
+            character_region.player, num_wave_cap_items_needed_for_wave
+        )
+        character_region.locations.append(wave_complete_location)
+
     return character_region
 
 
@@ -102,8 +124,5 @@ def create_loot_crate_group_region(
         )
 
     group_region.add_locations(group_locations, location_cls)
-    # group_region_rule = create_has_run_wins_rule(group_region.player, group.wins_to_unlock)
-    # parent_region.connect(group_region, name=group_region.name, rule=group_region_rule)
-    # regions.append(group_region)
 
     return group_region