+++ apworld/brotato/test/test_regions.py
@@ -15,7 +15,9 @@ from ..constants import (
 )
 from ..items import ItemName
 from ..loot_crates import BrotatoLootCrateGroup
+from ..options import NumWaveCaps
 from ..regions import create_character_region, create_loot_crate_group_region, create_regions
+from ..wave_caps import get_wave_cap_info
 from . import BrotatoTestBase
 
 
@@ -26,6 +28,7 @@ class TestBrotatoRegions(WorldTestBase):
         self.multiworld.game[1] = "Brotato"
         self.multiworld.player_name = {self.player: "Tester"}
         self.multiworld.worlds[1] = BrotatoWorld(self.multiworld, 1)
+        self.default_wave_access: dict[int, int] = get_wave_cap_info(NumWaveCaps(1))[1]
 
     def _create_region(self, name: str) -> Region:
         """Region factory to pass to the region creation functions."""
@@ -43,14 +46,14 @@ class TestBrotatoCharacterRegions(TestBrotatoRegions):
             WAVE_COMPLETE_LOCATION_TEMPLATE.format(char="Crazy", wave=15),
             WAVE_COMPLETE_LOCATION_TEMPLATE.format(char="Crazy", wave=20),
         ]
-        region = create_character_region(self._create_region, "Crazy", waves_with_checks)
+        region = create_character_region(self._create_region, "Crazy", waves_with_checks, self.default_wave_access)
         region_location_names = [loc.name for loc in region.locations]
 
-        self.assertListEqual(region_location_names, expected_location_names)
+        self.assertSequenceEqual(region_location_names, expected_location_names)
 
     def test_create_character_region_invalid_character_fails(self):
         with self.assertRaises(KeyError):
-            create_character_region(self._create_region, "Ironclad", [3, 6, 9, 12, 15, 18])
+            create_character_region(self._create_region, "Ironclad", [3, 6, 9, 12, 15, 18], self.default_wave_access)
 
     def test_create_character_region_invalid_waves_with_checks_fails(self):
         """Check that we don't create a region with invalid wave complete locations.
@@ -64,8 +67,10 @@ class TestBrotatoCharacterRegions(TestBrotatoRegions):
                 f"Check that create_character_region fails when waves_with_checks={invalid_value}",
                 invalid_value=invalid_value,
             ):
+                # Run setup to clear to the location cache between subtests
+                self.setUp()
                 with self.assertRaises(ValueError):
-                    create_character_region(self._create_region, "Brawler", invalid_value)
+                    create_character_region(self._create_region, "Brawler", invalid_value, self.default_wave_access)
 
 
 class TestBrotatoLootCrateRegions(TestBrotatoRegions):
@@ -121,6 +126,7 @@ class TestBrotatoCreateRegions(TestBrotatoRegions):
             self._create_region,
             self.characters,
             self.waves_with_checks,
+            self.default_wave_access,
             self.common_loot_crate_groups,
             self.legendary_loot_crate_groups,
         )
@@ -173,7 +179,10 @@ class TestBrotatoCreateRegions(TestBrotatoRegions):
 
 
 class TestBrotatoRegionAccessRules(BrotatoTestBase):
-    run_default_tests = False
+    @property
+    def run_default_tests(self) -> bool:
+        return False
+
     options: dict[str, Any] = {  # noqa: RUF012
         "num_victories": 10,
         "num_characters": 10,
@@ -191,6 +200,7 @@ class TestBrotatoRegionAccessRules(BrotatoTestBase):
             "Demon",
         ],
         "waves_per_drop": 4,
+        "num_wave_caps": 4,
         "num_common_crate_drops": 25,
         "num_common_crate_drop_groups": 5,
         "num_legendary_crate_drops": 5,
@@ -234,3 +244,48 @@ class TestBrotatoRegionAccessRules(BrotatoTestBase):
                 self.assertTrue(self.multiworld.state.can_reach_region(region_name, self.player))
             else:
                 self.assertAccessDependency(region_locations, [[char]])
+
+    def test_run_won_locations_have_correct_access_rules(self):
+        """Check that the run won locations require all wave cap increases to reach."""
+        characters = self.options["include_base_game_characters"]
+        for char in characters:
+            with self.subTest(character=char):
+                run_won_location_name = RUN_COMPLETE_LOCATION_TEMPLATE.format(char=char)
+
+                # Our test options say there are 3 cap increase items total
+                expected_items = [ItemName.PROGRESSIVE_WAVE_CAP_INCREASE.value] * 3
+                if char in self.world._starting_characters:
+                    expected_items.append(char)
+                self.assertAccessDependency([run_won_location_name], [expected_items], only_check_listed=True)
+
+    def test_wave_complete_locations_have_correct_access_rules(self):
+        """Check that wave complete locations are only reachable if the player has enough
+        Progressive Wave Cap Increase items.
+        """
+        characters = self.options["include_base_game_characters"]
+        for char in characters:
+            region_name = CHARACTER_REGION_TEMPLATE.format(char=char)
+            region = self.multiworld.regions.region_cache[self.player][region_name]
+            region_location_names = [loc.name for loc in region.locations]
+            expected_wave_location_access: dict[int, int] = {
+                4: 0,
+                8: 1,
+                12: 2,
+                16: 3,
+                20: 3,
+            }
+
+            for wave, expected_wave_cap_increases_needed in expected_wave_location_access.items():
+                with self.subTest(character=char, wave=wave):
+                    expected_items = [
+                        *[ItemName.PROGRESSIVE_WAVE_CAP_INCREASE.value] * expected_wave_cap_increases_needed,
+                    ]
+                    if char not in self.world._starting_characters:
+                        expected_items.append(char)
+                    location_name = WAVE_COMPLETE_LOCATION_TEMPLATE.format(char=char, wave=wave)
+                    # Sanity check that the location is defined
+                    self.assertIn(location_name, region_location_names)
+                    if expected_items:
+                        self.assertAccessDependency([location_name], [expected_items], only_check_listed=True)
+                    else:
+                        self.assertTrue(self.can_reach_location(location_name))