+++ CHANGES.md
@@ -7,6 +7,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- New Feature: Progressive Wave Caps
+  - The game now has a cap on the highest wave that can be played from the start.
+  - A new item, "Progressive Wave Cap Increase", has been added which, when received,
+    increases the maximum playable wave.
+  - The logic for Wave Completed and Run Won locations now requires that the wave cap be
+    greater than the wave number or 20, respectively.
+  - Controlled by a new option, `wave_caps`, which sets the number Progressive Wave Cap
+    Increase items and the increase given by each one.
+  - There are predefined values for `wave_caps`:
+    - `one`: One wave cap, all 20 waves are reachable from the start
+      (matches previous apworld behavior).\
+    - `two`: Waves 1-10 are playable from the start, waves 11-20 require a cap increase.
+    - `four`: Waves 1-5 are playable from the start, 6-10 require a single increase,
+      11-15 require two, 16-20 require three.
+    - `ten`: Waves 1-2 are playable from the start, 3-4 require one increase, etc.
+    - `twenty`: Every wave past wave 1 requires a cap increase.
+
+
 ### Changed
 
 - Reduced the maximum values of some options that were unfun at larger values:
@@ -17,6 +36,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   thresholds are evenly distributed over the number of wins needed to goal.
   - Before, the last group could be unlocked much earlier in logic if the number of
     groups did not evenly divide into the number of wins needed.
+- The progress UI in the Archipelago menu in the mod now shows the wave cap and number
+  of cap increases received.
 
 
 ## [0.15.2] - 2026-08-14